tests/requirements-test.txt
# Testing Requirements for E-Commerce Analytics Project
# Install with: pip install -r requirements-test.txt

# Core testing framework
pytest==7.4.3
pytest-cov==4.1.0
pytest-xdist==3.5.0  # For parallel test execution

# Test utilities
pytest-mock==3.12.0
pytest-timeout==2.2.0
hypothesis==6.92.1  # For property-based testing (optional)

# Core project dependencies (should match main requirements.txt)
pandas==2.1.4
numpy==1.26.2
pyyaml==6.0.1
pandera==0.17.2

# Additional utilities for testing
coverage==7.3.4
